Retrieval/requirements.txt
cffi
bs4
farm-haystack==1.2.0
sacremoses
faiss-gpu>=1.6.3,<2
sqlalchemy>=1.4.2,<2
sqlalchemy_utils
textblob
json5
requests
pydantic
tqdm
importlib-metadata
nltk
pandas
dill
fastapi
uvicorn
protobuf==3.20.*
